# Conflicts:
#	src/code/Manager.java

+++ src/code/Datenbank.java
@@ -13,6 +13,8 @@ public class Datenbank {
     private String mac;
     private String grund;
 
+    private boolean verbunden = true;
+
     private final Manager manager = new Manager();
     private final Dialogfenster dialogfenster = new Dialogfenster();
 
@@ -23,35 +25,37 @@ public class Datenbank {
 
     public void datenbankImportieren() { /* Importiert alle Datensätze, die aktuell in der Datenbank vorhanden sind, in die ArrayList */
         dialogfenster.verbindungWirdAufgebaut();
-        System.out.println("Connecting database...");
+
         Connection con = null; /* Verbindung zur Datenbank wird hergestellt */
         try {
             con = DriverManager.getConnection(url, username, password);
         } catch (CommunicationsException c) {
             dialogfenster.verbindungFehlgeschlagen();
+            verbunden = false;
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
-        System.out.println("Connected!");
 
-        Statement stmt1 = null;
-        try {
-            stmt1 = con.createStatement();
-            ResultSet rs1 = stmt1.executeQuery("SELECT * FROM table2");
+        if (verbunden) {
+            Statement stmt1 = null;
+            try {
+                stmt1 = con.createStatement();
+                ResultSet rs1 = stmt1.executeQuery("SELECT * FROM table2");
 
-            int anzahl = 0;
-            while (rs1.next()) { /* Cursor der Datenbank wechselt mit jedem rs1.next()-Aufruf in die nächste Zeile */
-                kursstufe = rs1.getString("Kursstufe");
-                nachname = rs1.getString("Nachname");
-                vorname = rs1.getString("Vorname");
-                mac = rs1.getString("MAC");
-                grund = rs1.getString("Grund");
-                manager.ergaenze(kursstufe, nachname, vorname, mac, grund); /* Aus jeder Zeile wird ein Datensatz-Objekt erstellt und in die Liste eingefügt */
-                anzahl++;
+                int anzahl = 0;
+                while (rs1.next()) { /* Cursor der Datenbank wechselt mit jedem rs1.next()-Aufruf in die nächste Zeile */
+                    kursstufe = rs1.getString("Kursstufe");
+                    nachname = rs1.getString("Nachname");
+                    vorname = rs1.getString("Vorname");
+                    mac = rs1.getString("MAC");
+                    grund = rs1.getString("Grund");
+                    manager.ergaenze(kursstufe, nachname, vorname, mac, grund); /* Aus jeder Zeile wird ein Datensatz-Objekt erstellt und in die Liste eingefügt */
+                    anzahl++;
+                }
+                dialogfenster.DatenbankImportiert(anzahl);
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
             }
-            dialogfenster.DatenbankImportiert(anzahl);
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
         }
     }
 
@@ -60,16 +64,18 @@ public class Datenbank {
             dialogfenster.datenbankListeLeer();
         } else {
             dialogfenster.verbindungWirdAufgebaut();
-            System.out.println("Connecting database...");
             Connection con = null; /* Verbindung zur Datenbank wird hergestellt */
             try {
                 con = DriverManager.getConnection(url, username, password);
+            } catch (CommunicationsException c) {
+                dialogfenster.verbindungFehlgeschlagen();
+                verbunden = false;
             } catch (SQLException throwables) {
                 throwables.printStackTrace();
             }
-            System.out.println("Connected!");
 
             int len = liste.size();
+            boolean hochgeladen = false;
             for (int i = 0; i < len; i++) {
                 Datensatz aktDatensatz = liste.get(i);
                 kursstufe = aktDatensatz.getKursstufe();
@@ -78,18 +84,22 @@ public class Datenbank {
                 mac = aktDatensatz.getMac();
                 grund = aktDatensatz.getGrund();
 
-                Statement stmt1 = null;
-                try {
-                    stmt1 = con.createStatement();
-                    stmt1.executeUpdate("INSERT INTO Table2 (Kursstufe, Nachname, Vorname, MAC, Grund) VALUES ('" + kursstufe + "','" + nachname + "','" + vorname + "','" + mac + "','" + grund + "')");
-                } catch (SQLException throwables) {
-                    throwables.printStackTrace();
-                } catch (NullPointerException e) {
-                    e.printStackTrace();
-                    System.out.println("NullPointerException");
+                if (verbunden) {
+                    Statement stmt1 = null;
+                    try {
+                        stmt1 = con.createStatement();
+                        stmt1.executeUpdate("INSERT INTO Table2 (Kursstufe, Nachname, Vorname, MAC, Grund) VALUES ('" + kursstufe + "','" + nachname + "','" + vorname + "','" + mac + "','" + grund + "')");
+                        hochgeladen = true;
+                    } catch (NullPointerException e) {
+                        e.printStackTrace();
+                    } catch (SQLException throwables) {
+                        throwables.printStackTrace();
+                    }
                 }
             }
-            dialogfenster.DatenbankErgaenzt(len);
+            if (hochgeladen) {
+                dialogfenster.DatenbankErgaenzt(len);
+            }
         }
     }
 
@@ -99,19 +109,24 @@ public class Datenbank {
         Connection con = null; /* Verbindung zur Datenbank wird hergestellt */
         try {
             con = DriverManager.getConnection(url, username, password);
+        } catch (CommunicationsException c) {
+            dialogfenster.verbindungFehlgeschlagen();
+            verbunden = false;
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
         System.out.println("Connected!");
 
-        Statement stmt1 = null;
-        try {
-            stmt1 = con.createStatement();
-            stmt1.executeUpdate("TRUNCATE Table2");
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-        }
+        if (verbunden) {
+            Statement stmt1 = null;
+            try {
+                stmt1 = con.createStatement();
+                stmt1.executeUpdate("TRUNCATE Table2");
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
+            }
 
-        dialogfenster.DatenbankGeleert();
+            dialogfenster.DatenbankGeleert();
+        }
     }
 }
